add onlinejudge/utils.py to provide functions about sessions as a library
-   see https://github.com/kmyk/online-judge-tools/issues/480#issuecomment-565946620
